First-pass implementation of bot creation.
This is as bare as it gets, just something so I can put it on a linode and
get it producing. Hard-codes 12hour tweets, uses channels and goroutines
so I have NO IDEA whether this works for anything. But we'll see... if I'm
understanding how these work correctly, it's actually pretty hip.